add: bit declaration code
